+++ src/test/java/io/github/astrapi69/gradle/migration/data/JacksonTomlExtensions.java
@@ -1,0 +1,52 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2023 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.github.astrapi69.gradle.migration.data;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.List;
+import java.util.Map;
+
+import com.fasterxml.jackson.dataformat.toml.TomlMapper;
+
+public class JacksonTomlExtensions
+{
+
+	public static Map read(File tomlFile) throws IOException
+	{
+		TomlMapper mapper = new TomlMapper();
+		Map map = mapper.readValue(tomlFile, Map.class);
+		return map;
+	}
+
+
+	public static void writeValue(File tomlFile, Map<String, Map> libsVersionTomlMap)
+		throws IOException
+	{
+		TomlMapper mapper = new TomlMapper();
+		String valueAsString = mapper.writeValueAsString(libsVersionTomlMap);
+		mapper.writeValue(tomlFile, valueAsString);
+	}
+}